+++ eperd/sslgetcert.c
@@ -182,7 +182,7 @@ static void buf_add(struct buf *buf, const void *data, size_t len)
 	newbuf= malloc(maxsize);
 	if (!newbuf)
 	{
-		fprintf(stderr, "unable to allocate %ld bytes\n", maxsize);
+		fprintf(stderr, "unable to allocate %ld bytes\n", (long)maxsize);
 		exit(1);
 	}
